split code on mygraph.

+++ nvtc/MyGraph.h
@@ -13,7 +13,7 @@
 #include <mutex>
 using namespace std;
 
-int get_split_v2(int64_t* offset, int nodeid_max, int split_num, int64_t*& out);
+int64_t get_split_v2(int64_t* offset, int nodeid_max, int split_num, int64_t*& out);
 
 class MyGraph{
 	public:
